% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness.R
\name{completness}
\alias{completness}
\title{Completeness analysis}
\usage{
completness(nObs)
}
\arguments{
\item{nObs}{an object of any class (mainly resulting from \code{summariseBirds} or
\code{exportBirds} with  the number of observations, or visits in your desired analysis unit.}

\item{nSpp}{the number of unique species observed}

\item{h}{the half ignorance parameter value.}
}
\value{
a \code{data.frame} with ignorance scores
}
\description{
Lorem ipsum
}
\examples{
\donttest{
OB <- organizeBirds(bombusObsShort, sppCol = "scientificName", simplifySppName = TRUE)
grid <- makeGrid(searchPolygon, gridSize = 10)
SB <- summariseBirds(OB, grid=grid)
ignorance <- exposeIgnorance(nObs=SB$spatial@data$nObs)
}
}
\references{
La Sorte & Somveille (2019) < https://doi.org/10.1111/ecog.04632>
}
\seealso{
\code{\link{summarizeBirds}}, \code{\link{exportBirds}}
}
