% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBirds.R
\name{countsYearMonth}
\alias{countsYearMonth}
\title{Internal function for getSpatiotemporal()}
\usage{
countsYearMonth(x, yearsAll, visitCol)
}
\description{
An internal function to be applied in a lapply
}
\keyword{internal}
