% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportBirds.R
\name{deconstructOverlay}
\alias{deconstructOverlay}
\title{Deconstruction of the overlay}
\usage{
deconstructOverlay(overlay, visitCol)
}
\arguments{
\item{overlay}{A list with an element by grid cell}

\item{visitCol}{A character string specifying the columns that identify a visit.}
}
\value{
a long data.frame
}
\description{
This takes the overlay element of a summarisedBirds, reconstruct it as a
A long data.frame and removes duplicate visits in other grid cells.
}
\keyword{internal}
