% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeBirds.R
\name{findCols}
\alias{findCols}
\title{Find a text among column names}
\usage{
findCols(pattern, df, exact = FALSE, value = TRUE)
}
\arguments{
\item{pattern}{character string containing a string or regular expression
to be matched in the given character vector}

\item{df}{A data.frame among whichs column names to look for the pattern}

\item{exact}{A logical flag whether to return all matches or just that text.
This is always case insensitive.}

\item{value}{A logical flag (to be passed to grep()). If FALSE, a vector
containing the (integer) indices of the matches determined by grep is returned,
and if TRUE, a vector containing the matching elements themselves is returned.}
}
\value{
A vector with the column names that match the pattern
}
\description{
Finds a text among column names and return it as it is
}
\keyword{internal}
