% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focalSpecies.R
\name{focalSpSummary}
\alias{focalSpSummary}
\title{A function to find a species among all visits}
\usage{
focalSpSummary(x, focalSp = NULL)
}
\arguments{
\item{x}{an object of class \sQuote{SummarizeBirds}.}

\item{focalSp}{the focal sp to look for.}
}
\value{
a \code{data.frame} with summary data for the focal species
}
\description{
A function to find a species among all visits.
}
\examples{
OB <- organizeBirds(bryophytaObs, sppCol = "scientificName", simplifySppName = TRUE)
grid <- makeGrid(searchPolygon, gridSize = 10)
SB <- summariseBirds(OB, grid=grid)
allSpp <- listSpecies(SB)
focal<-"Zygodon viridissimus Bridel, 1826"
focalSpSummary(SB, focalSp=focal)
}
\seealso{
\code{\link{summarizeBirds}}, \code{\link{exportBirds}}
}
