% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeBirds.R
\name{getGridIDs}
\alias{getGridIDs}
\title{Create unique IDs based on a grid}
\usage{
getGridIDs(x, grid)
}
\arguments{
\item{x}{a ... .
be a SpatialPointsDataFrame (useful when inside 'organizeBirds()') or an OrganizedBirds.}

\item{grid}{A a SpatialPolygon object (a grid is expected) defining the
maximum extent of visits effort.}
}
\value{
A vector of the same length as the number of rows (observations) as x
  with a unique number corresponding to the grid's ID.
}
\description{
Takes a dataframe and a grid .
}
\examples{


}
