% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBirds.R
\name{getSpatioTemporal}
\alias{getSpatioTemporal}
\title{Collect spatiotemporal data}
\usage{
getSpatioTemporal(birdOverlay, visitCol = NULL)
}
\arguments{
\item{visitCol}{A character string for specifying the columns that identify a visit.}

\item{birdData}{An OrganizedBirds object, which is the input to the \code{\link{overlayBirds}}-function.}
}
\value{
A list with two arrays
}
\description{
An internal function. Takes the dataframe resulting from the spatial overlay
for a specific grid and summarises basic variables as a list of arrays
}
\keyword{internal}
