% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBirds.R
\name{getTemporal}
\alias{getTemporal}
\title{Create spillover overlay for specific grid}
\usage{
getTemporal(birdData, visitCol = NULL)
}
\arguments{
\item{birdData}{An OrganizedBirds object, which is the input to the \code{\link{overlayBirds}}-function.}

\item{visitCol}{A character string for specifying the columns that identify a visit.}

\item{data}{A dataframe for a specific grid without spillover.}
}
\value{
The overlay dataframe for a specific grid, including the spillover visits.
}
\description{
An internal function. Takes the dataframe resulting from the spatial overlay
for a specific grid and searches for spillover visits (i.e. visits that are split by grid cell boundaries).
}
\keyword{internal}
