% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUTM.r
\name{getUTMproj}
\alias{getUTMproj}
\title{A wrapper around getUTMzone and produce a crs object string}
\usage{
getUTMproj(x)
}
\arguments{
\item{x}{an object of class \sQuote{OrganizedBirds}, \sQuote{sf} or \sQuote{SpatialPointsDataFrame}}
}
\value{
a EPSG integer code for an appropriate UTM zone
}
\description{
A wrapper around getUTMzone and produce a crs object string
}
\examples{
OB <- organizeBirds(bombusObs)
getUTMproj(OB)
}
