% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUTM.r
\name{getUTMzone}
\alias{getUTMzone}
\title{internal function. Depends on the polygon object utmZones}
\usage{
getUTMzone(points)
}
\arguments{
\item{points}{a set of points}
}
\description{
internal function. Depends on the polygon object utmZones
}
\keyword{internal}
