% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gotaland}
\alias{gotaland}
\title{Götaland, Sweden.}
\format{A SpatialPolygonsDataFrame with 1 polygon with CRS(\dQuote{+init=epsg:4326})}
\usage{
gotaland
}
\description{
A polygon describing the territorial contour the collective province of Götaland, Sweden. Includes
holes for water bodies.
}
\keyword{datasets}
