% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGrid.r
\name{gridAsString}
\alias{gridAsString}
\title{Convert a grid into a web query string.}
\usage{
gridAsString(grid)
}
\arguments{
\item{grid}{an object of class \sQuote{SpatialPolygon-class} or
\sQuote{SpatialPolygonDataFrame-class}.}
}
\value{
a character string with coordinates separated by \dQuote{\%20} and pairs by \dQuote{,}.
}
\description{
Converts a grid (or any SpatialPolygon for that matter) into a web query string.
}
