% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridGotaland}
\alias{gridGotaland}
\title{Götalands grid, Sweden.}
\format{
A sf with 1 polygon with CRS(epsg:4326)
}
\usage{
gridGotaland
}
\description{
A polygon with a sample hexagonal grid over Götaland, Sweden.
}
\keyword{datasets}
