% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayBirds.R
\name{includeSpillover}
\alias{includeSpillover}
\title{Include spillover observations}
\usage{
includeSpillover(x, birdData, visitCol)
}
\arguments{
\item{x}{An observationsInGrid list from \code{\link{overlayBirds}}. This
is also what will be modified and become the output.}

\item{birdData}{An OrganizedBirds object, which is the input to the \code{\link{overlayBirds}}-function.}

\item{visitCol}{A character string for specifying the columns that identify a visit.}
}
\value{
A modified observationsInGrid list
}
\description{
An internal function to include spillover observations to the gridcells.
}
\keyword{internal}
