% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayBirds.R
\name{includeUniqueSpillover}
\alias{includeUniqueSpillover}
\title{Include unique spillover observations}
\usage{
includeUniqueSpillover(birdData, grid, visitCol)
}
\arguments{
\item{birdData}{An OrganizedBirds object, which is the input to the \code{\link{overlayBirds}}-function.}

\item{grid}{A SpatialPolygonsDataFrame object of the grid over the study area}

\item{visitCol}{A character string for specifying the columns that identify a visit.}
}
\value{
A  ObservationsInGrid list
}
\description{
An internal function to include spillover observations to the gridcells.
This function moves spillover observations to the gridcells were the focal visit has the most observations (or at random if ties)
}
\keyword{internal}
