% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreVisits.r
\name{lapplyVisits}
\alias{lapplyVisits}
\title{An internal function to explore the definition of field visits}
\usage{
lapplyVisits(x, dat, visitCol, spdf, minPts)
}
\arguments{
\item{x}{an subset object to pass to apply}

\item{dat}{an object of class \sQuote{OrganizedBirds} (organised BIRDS Spatial data frame).}

\item{visitCol}{name of the column for the visits UID.}

\item{spdf}{spatial object of dat.}

\item{minPts}{minimum number of points, default to 3.}
}
\value{
a \code{data.frame} with summarized data per visit:
\itemize{
  \item \dQuote{effortDiam}: the 2 times the maximum of the distances between
  the centroid of all observation points and any individual observation.
  \item \dQuote{medianDist}: the median (Q2) of the distances between the
  centroid and the observations, in meters.
  \item \dQuote{iqrDist}: the interquartile range of the distances between the
   centroid the observations, in meters.
  \item \dQuote{nUniqueLoc}: the number of unique combination of coordinates (locations).
  \item \dQuote{nClusters}: the number of clusters defined by the DBSCAN algorithm,
  a minimum of 3 observations per cluster within the median distance between
  all observations. If the number of clusters is = 0 means that there are at
  least 3 unique locations but observations are too spread and no cluster was
  found. If the number of unique locations is less than 3, observations are
  considered as a single cluster without outliers.
  \item \dQuote{nOutliers}: the number of observations whose distance to any
  cluster is beyond the median distance between all observations.
}
}
\description{
An internal function to explore the definition of field visits
}
\seealso{
\code{\link{createVisits}}, \code{\link{organiseBirds}}
}
\keyword{internal}
