% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focalSpecies.R
\name{listSpecies}
\alias{listSpecies}
\title{Five functions extracting summaries for species:
1) A function to list all species from all visits
2) A function to find records of a species among all visits
3) A function to summarise records for a species
4) A function to list all species and count number of visits
5) A function to create a community matrix}
\usage{
listSpecies(x)
}
\arguments{
\item{x}{an object of class \sQuote{OrganizedBirds} or \sQuote{SummarizeBirds}.}
}
\value{
a \code{vector} with all species names in the data set
}
\description{
A function to list all species from all visits
}
\details{
A function to list all species from all visits (total species list for the data set).
}
\examples{
OB <- organizeBirds(bryophytaObs, sppCol = "scientificName", simplifySppName = TRUE)
allSpp <- listSpecies(OB)
}
\seealso{
\code{\link{summarizeBirds}}, \code{\link{exportBirds}}
}
