% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGrid.r
\name{makeCircle}
\alias{makeCircle}
\title{Create the minimum circle containing the points}
\usage{
makeCircle(spdf, projCRS = NULL)
}
\arguments{
\item{spdf}{an object of class \sQuote{SpatialPointsDataFrame} with defined CRS.}

\item{projCRS}{a proj4-type string defining a projected CRS. It is very important
that the selected CRS is accurate in the study area.}
}
\value{
a polygon object of class \sQuote{SpatialPolygon} with geodesic
coordinates in WGS84 (ESPG:4326).
}
\description{
This funciton is based on the function shotGroups::getMinCircle() that uses
the Skyum algorithm based on the convex hull.
}
