% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observationIndex.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
a positive numeric vector
}
\description{
This function nomalizes a vector to a 0-1 range
}
\keyword{internal}
