% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeBirds.R
\name{obsData.OrganizedBirds}
\alias{obsData.OrganizedBirds}
\title{Extract observation data}
\usage{
\method{obsData}{OrganizedBirds}(x)
}
\arguments{
\item{x}{An OrganizedBirds-object}
}
\description{
Extract observation data
}
