% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeBirds.R
\name{obsData}
\alias{obsData}
\title{Extract observation data}
\usage{
obsData(x, ...)
}
\arguments{
\item{x}{An OrganizedBirds-object}

\item{...}{additional arguments for \code{\link{obsData}}}
}
\value{
A dataframe
}
\description{
Extract the observation data from a OrganizedBirds-object
}
\examples{
ob<-organizeBirds(bombusObs)
obsData(ob)
}
