% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observationIndex.R
\name{obsIndexSpatial}
\alias{obsIndexSpatial}
\title{Relative observation index (Spatial)}
\usage{
obsIndexSpatial(x, focalSp = NULL, visits = TRUE, fs.rm = TRUE, norm = TRUE)
}
\arguments{
\item{x}{an object of class \sQuote{SummarizeBirds}.}

\item{focalSp}{the focal sp to look for}

\item{visits}{if TRUE (default) the observation index is calculated over number
of visits, else uses the number of observations.}

\item{fs.rm}{if TRUE, assumes that the observations for the focal species are
included in 'group' and will remove them}

\item{norm}{if TRUE, the result is normalized to a 0-1 range}
}
\value{
A spatial object
}
\description{
This function extracts the proportion of visits (or observations) detecting a
focal species to all visits (or observations) over space.
}
\keyword{internal}
