% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observationIndex.R
\name{obsIndexTemporal}
\alias{obsIndexTemporal}
\title{Relative observation index (Temporal)}
\usage{
obsIndexTemporal(x, timeRes, focalSp = NULL)
}
\arguments{
\item{x}{an object of class \sQuote{SummarizeBirds}.}

\item{timeRes}{the time resolution:  \code{"Yearly", "Monthly"} or \code{"Daily"}}

\item{focalSp}{the focal sp to look for.}
}
\value{
An xts timeseries
}
\description{
This function extracts the proportion of observations for a focal species to
the all observations (including the focal species) over time.
}
