% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeBirds.R
\name{organizeBirds}
\alias{organizeBirds}
\alias{organiseBirds}
\title{Organize a dataframe to a usable format}
\usage{
organizeBirds(x, sppCol = "scientificName", timeCol = c(Year = "year",
  Month = "month", Day = "day"), visitsIdentifier = c("locality", "day",
  "month", "year", "recordedBy"), presenceCol = NULL,
  xyCols = c("decimalLongitude", "decimalLatitude"),
  dataCRS = "+init=epsg:4326", taxonRankCol = NULL,
  taxonRank = c("SPECIES", "SUBSPECIES", "VARIETY"),
  simplifySppName = FALSE)
}
\arguments{
\item{x}{A dataframe or a SpatialPointsDataFrame containing at least a
column for species name, one or several columns for date of observation, one or
several columns for identifying a visit and, if it is not spatial, coordinate
columns.}

\item{sppCol}{A character string with the column name for the column for the
species names. Default is the Darwin Core standard name \code{"scientificName"}.}

\item{timeCol}{A character vector (named or not) of the names for the
column(s) which is holding the observation dates. Default is the Darwin Core
standard column names \code{c("Year"="year", "Month"="month", "Day"="day")}.}

\item{visitsIdentifier}{A character vector of the names for the columns that
are holding the information that identifies a visit. Default is the Darwin
Core standard column names \code{c("locality", "day", "month", "year",
"recordedBy")}.}

\item{presenceCol}{A character string with the column name for the column for the
precense status. Default is \code{NULL}.}

\item{xyCols}{A character vector of the names for the columns that are holding
the coordinates for the observations. The order should be longitude(x),
latitude(y). Default is the Darwin Core standard column names
\code{c("decimalLongitude", "decimalLatitude")}. Only applicable to non-
spatial dataframes.}

\item{dataCRS}{A character string for the dataframe CRS (Coordinate Reference
System). Default is \code{"+init=epsg:4326"}, which is WGS 84. This is only
applicable to non-spatial dataframes, since a spatial dataframes already
should have this information.}

\item{taxonRankCol}{the name of the column containing the taxonomic rank for the observation.
That is the minimum taxonomic identification level.}

\item{taxonRank}{a string or vector of strings containing the taxonomic ranks to keep.
Only evaluated if taxonRankCol is not \code{NULL}}

\item{simplifySppName}{wheter to remove everything else that is not the species name (authors, years and intraspecific epithets). Default set to FALSE}
}
\value{
a `SpatialPointsDataFrame` wrapped into an object of class OrganizedBirds, with additional attributes.
}
\description{
Takes a dataframe with reported species observations and reformats it, using 
visit identifiers, to an OrganizedBirds-class that can be used in further 
analyses with the BIRDS-package.
}
\details{
An OrganizedBirds-class is essentially a list containing one element, a
SpatialPointsDataFrame. This SpatialPointsDataFrame has its data formatted in
a way that the other functions in the BIRDS-package can use further on. It
also has the attribute \code{"visitCol"}, which indicates which column in the
dataframe holds the visit identifier. The visit identifier is created by
the function \code{\link{createVisits}}, which creates a unique id for each
combination of the values in the defined columns.

The variable \code{timeCol} can be formatted differently. If the variable is a
named vector with the names "Year", "Month" and "Day" (letter capitalization
does not matter) it will use the variable named year as the year column and so
on. Otherwise it will use the first variable as year, the second as month and
the third as day, if there is a vector of length three or more. If the vector
is of only length one it will interpret the column as a date column formatted
as "yyyy-mm-dd".
}
\examples{
organizeBirds(bombusObs)
}
\seealso{
\code{\link{createVisits}} to create unique visits IDs,
 \code{\link{visits}} to get or set the visit IDs to this class,
 \code{\link{obsData}} to retrieve the dataframe from this class.
}
