% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayBirds.R
\name{overlayBirds}
\alias{overlayBirds}
\alias{overlayBirds.OrganizedBirds}
\title{Overlay BIRDS object and grid}
\usage{
overlayBirds(x, grid, spillOver = TRUE)

\method{overlayBirds}{OrganizedBirds}(x, grid, spillOver = TRUE)
}
\arguments{
\item{x}{An OrganizedBirds object}

\item{grid}{A SpatialPolygonsDataFrame object of the grid in the study area}

\item{spillOver}{Specifies if the function should search for observations belonging to the same
visit (identified by visit UID) but fall outside of the grid cell. Default is \code{TRUE}.}
}
\value{
The output is a OverlaidBirds-class object, which is a list
  containing three objects;
  \describe{
    \item{\code{observationsInGrid}}{Is
      basically the data in the OrganizedBirds object split by each grid cell
      (\emph{n.b.} the use of \code{spillOver = TRUE} discussed under "Usage")}
    \item{\code{grid}}{The SpatialPolygonsDataFrame from the input, but cleared
      of data to not waste unnecessary memory}
   \item{\code{nonEmptyGridCells}}{An integer vector of which grid cells that have observations}
  }

The output is a OverlaidBirds-class object, which is a list
  containing three objects;
  \describe{
    \item{\code{observationsInGrid}}{Is
      basically the data in the OrganizedBirds object split by each grid cell
      (\emph{n.b.} the use of \code{spillOver = TRUE} discussed under "Usage")}
    \item{\code{grid}}{The SpatialPolygonsDataFrame from the input, but cleared
      of data to not waste unnecessary memory}
   \item{\code{nonEmptyGridCells}}{An integer vector of which grid cells that have observations}
  }
}
\description{
Make an overlay for an OrganizedBirds object and a grid to identify which
observations that fall into which grid cell.

This function takes an OrganizedBirds object, created by
\code{\link{organizeBirds}}, and a polygon grid-layer, which could be created
by \code{\link{makeGrid}} and splits the visits in the OrganizedBirds
(i.e. data belonging to identified visits) to each grid cell.
If \code{spillOver = FALSE} the splitting is done spatially according to the overlay of observations
and grid cells, without further consideration of coherence for visits (visit UID).
If \code{spillOver = TRUE} the splitting will be done spatially in a first step,
but in the next step for each grid cell the function searches for other observations belonging
to the visits inside the gridcell (identified by visit UID) in the entire dataset and includes
all such observations found to the observations in the grid cell (i.e. keeping visits coherent).
}
\details{
This function takes an OrganizedBirds object, created by
\code{\link{organizeBirds}}, and a polygon grid-layer, which could be created
by \code{\link{makeGrid}} and splits the visits in the OrganizedBirds
(i.e. data belonging to identified visits) to each grid cell.
If \code{spillOver = FALSE} the splitting is done spatially according to the overlay of observations
and grid cells, without further consideration of coherence for visits (visit UID).
If \code{spillOver = TRUE} the splitting will be done spatially in a first step,
but in the next step for each grid cell the function searches for other observations belonging
to the visits inside the gridcell (identified by visit UID) in the entire dataset and includes
all such observations found to the observations in the grid cell (i.e. keeping visits coherent).
}
\examples{
ob<-organizeBirds(bombusObs)
grid <- makeGrid(gotaland, gridSize = 10)
overlayBirds(ob, grid)
ob<-organizeBirds(bombusObs)
grid <- makeGrid(gotaland, gridSize = 10)
overlayBirds(ob, grid)
}
