% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relObsTrends.R
\name{relObsTrends}
\alias{relObsTrends}
\title{Relative Observation Trends}
\usage{
relObsTrends(x, timeRes, focalSp)
}
\arguments{
\item{x}{an object of class \sQuote{SummarizeBirds}.}

\item{timeRes}{the time resolution:  \code{"Yearly", "Monthly"} or \code{"Daily"}}

\item{focalSp}{the focal sp to look for.}
}
\value{
A timeseri
}
\description{
Relative Observation Trends
}
