% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportBirds.R
\name{removeInexDays}
\alias{removeInexDays}
\title{A funciton to remove inexistent combination of days like april 31. X is the result of group by with dates as factors}
\usage{
removeInexDays(x)
}
\description{
A funciton to remove inexistent combination of days like april 31. X is the result of group by with dates as factors
}
\keyword{internal}
