% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGrid.r
\name{renameGrid}
\alias{renameGrid}
\title{Rename the cells in a grid}
\usage{
renameGrid(grid)
}
\arguments{
\item{grid}{an object of class \sQuote{SpatialPolygon-class} or
\sQuote{SpatialPolygonDataFrame-class}.}
}
\value{
the same input object with known names
}
\description{
Takes a SpatialPolygon* and renames it to "ID1":"IDn".
}
\keyword{internal}
