% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SB}
\alias{SB}
\title{A simple summarisedBirds object.}
\format{
An object of class summarisedBirds
}
\usage{
SB
}
\description{
A summarisedBirds object based on bombusObsShort for faster examples
grid <- makeGrid(searchPolygon, gridSize = 10)
SB <- summarizeBirds(organizeBirds(bombusObsShort), grid=grid)
}
\note{
some special characters have been transformed to ASCII to comply with CRAN requirements
}
\keyword{datasets}
