% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{searchPolygon}
\alias{searchPolygon}
\title{Searching polygon covering Götaland, Sweden.}
\format{
A sf with 1 polygon with CRS(epsg:4326)
}
\usage{
searchPolygon
}
\description{
An arbitrary polygon covering a portion of Sweden, used as search parameter.
}
\keyword{datasets}
