% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simpleSB}
\alias{simpleSB}
\title{A simple empty summarisedBirds object.}
\format{An object of class summarisedBirds}
\usage{
simpleSB
}
\description{
An empty summarisedBirds object used to dinamically test for validity
of export parameter combinations \code{exportBirds()} in the sister package shinyBirds
}
\keyword{datasets}
