% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeBirds.R
\name{simplifySpp}
\alias{simplifySpp}
\title{Simplify species names}
\usage{
simplifySpp(df, sppCol)
}
\arguments{
\item{df}{A dataframe with at least the columns specified in cols}

\item{sppCol}{A character vector with the column names for the species names.}
}
\value{
A vector vith data.frame with a scientific names up to specific names
}
\description{
Removes infraspecific epithets, authors and years from scientific names
}
