% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBirds.R
\name{summariseBirds}
\alias{summariseBirds}
\alias{summarizeBirds}
\title{Summarize the OrganizedBirds}
\usage{
summariseBirds(x, ...)
}
\arguments{
\item{x}{An OrgnanizedBirds-object created by \code{\link{organizeBirds}}}

\item{...}{additional parameters for \code{\link{summarizeBirds}}}
}
\description{
Takes a OrganizedBirds-object and a SpatialPolygons*-grid and summarizes it
in spatial and  temporal dimensions.
}
