% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBirds.R
\name{summarizeBirds.OrganizedBirds}
\alias{summarizeBirds.OrganizedBirds}
\title{Summarize the OrganizedBirds}
\usage{
\method{summarizeBirds}{OrganizedBirds}(x, grid, spillOver = TRUE)
}
\arguments{
\item{x}{An OrgnanizedBirds-object created by \code{\link{organizeBirds}}}

\item{grid}{A SpatialPolygons or SpatialPolygonsDataFrame-object whith
grid cells for the study area. This variable is optional and can be set to
NULL, which will treat the area for all observations as one single
grid cell.}

\item{spillOver}{Specifies if the function should search for observations done during
the same visit (identified by visit UID) but that fall outside the grid cell. Default is \code{TRUE}.}
}
\description{
Takes a OrganizedBirds-object and a SpatialPolygons*-grid and summarizes it
in spatial and  temporal dimensions.
}
