% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeBirds.R
\name{createVisits}
\alias{createVisits}
\alias{visits}
\alias{\sQuote{visits<-}}
\alias{visits<-}
\title{Create unique visits IDs}
\usage{
createVisits(x, columns = c("locality", "day", "month", "year",
  "recordedBy"))

visits(x, name=NULL)

visits(x, name = NULL, useAsDefault = TRUE) <- value
}
\arguments{
\item{x}{a dataframe or OrganizedBirds-class including at least the
specified columns that are used to identify a visit.}

\item{columns}{A vector with the names of the columns that are used to identify a visit.
Default is the Darwin Core variables \code{c("locality", "day", "month",
"year", "recordedBy")}.}

\item{name}{The name of the visit column. Default is \code{NULL}, which will
get/write to the predefined visit column (\code{visitUID}).}

\item{useAsDefault}{Specifies if the defined column in \code{name} should be used as
the default column for the visits in further analysis. If name is
\code{NULL} and \code{useAsDefault = TRUE}, \code{value} will be written to
column (\code{visitUID}) and that column will be set to default.}

\item{value}{the value to assign}

\item{x}{An OrganizedBirds-object}
}
\value{
A vector of the same length as the number of rows as the dataframe
  with a unique number for each combination of the values in the specified
  columns.
}
\description{
Takes a dataframe and a vector of column names and classifies each row of the
dataframe based on the combination of values in the specified columns.

Gets or sets the visits identifier for a OrganizedBirds-class.
}
\details{
What a visit should be is not always clearly defined and extractable in a
dataset. A reasonable assumption is that a visit could be identified from the records
made by one person on a certain day and at a specific location or site. The
default value for the variable column is therefore that a visit is identified
by the Darwin Core variables \code{c("locality", "day", "month", "year",
"recordedBy")}.
}
\examples{
OB <- organizeBirds(bombusObs)
tmp.vis <- createVisits(bombusObs, columns=c("locality", "day", "month", "year"))
visits(OB, name = "visNoRecorder", useAsDefault = TRUE) <- tmp.vis
ob<-organizeBirds(bombusObs)
attr(ob, "visitCol")
vis<-visits(ob)
tmp.vis <- createVisits(bombusObs, columns=c("locality", "day", "month", "year"))
visits(ob, name = "visNoRecorder", useAsDefault = TRUE) <- tmp.vis
vis2<-visits(ob)
attr(ob, "visitCol")
}
