% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayBirds.R
\name{whichNonEmpty}
\alias{whichNonEmpty}
\title{Which cells are not empty}
\usage{
whichNonEmpty(x)
}
\arguments{
\item{x}{A list.}
}
\value{
The a \code{vector} with list elements IDs.
}
\description{
This function returns the ID of those cells that have some data. It is used to
speed up other functions
}
