% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bombusObsShort}
\alias{bombusObsShort}
\title{A short version of bombusObs dataset for faster examples.}
\format{
A data frame with 1,000 rows and 45 variables following DarwinCore standard \url{https://dwc.tdwg.org/}
}
\source{
\url{https://www.gbif.org/occurrence/download/0007731-190320150433242}
}
\usage{
bombusObsShort
}
\description{
A random sample of 1,000 observations for the genus \emph{Bombus} spp.
}
\note{
some special characters have been transformed to ASCII to comply with CRAN requirements
}
\seealso{
\code{\link{bombusObs}}
}
\keyword{datasets}
