% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposeIgnorance.R
\name{exposeIgnorance}
\alias{exposeIgnorance}
\title{A function to create Ignorance Scores}
\usage{
exposeIgnorance(nObs, nSpp = NULL, h = 1)
}
\arguments{
\item{nObs}{an object of any class (mainly resulting from \code{summariseBirds} or
\code{exportBirds} with  the number of observations, or visits in your desired analysis unit.}

\item{nSpp}{the number of unique species observed}

\item{h}{the half ignorance parameter value.}
}
\value{
a \code{data.frame} with ignorance scores
}
\description{
A function to create Ignorance Scores.
Ignorance scores are a proxy for the lack of sampling effort, computed by relating
a focal species’ occurrence to occurrences of all species in a reference species group.
Read more here: Ruete 2015 Biodiv Data J 3:e5361, doi:10.3897/BDJ.3.e5361
}
\examples{
OB <- organizeBirds(bryophytaObs, sppCol = "scientificName", simplifySppName = TRUE)
grid <- makeGrid(searchPolygon, gridSize = 10)
SB <- summariseBirds(OB, grid=grid)
ignorance <- exposeIgnorance(nObs=SB$spatial@data$nObs)
}
\seealso{
\code{\link{summarizeBirds}}, \code{\link{exportBirds}}
}
