% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observationIndex.R
\name{extractPresence}
\alias{extractPresence}
\title{Extract precense from SB}
\usage{
extractPresence(x)
}
\arguments{
\item{x}{an object of class \sQuote{SummarizeBirds}.}
}
\value{
returns x without observations where the value precense is < 1
}
\description{
This function removes abcense observations of the focal species
}
\keyword{internal}
