% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeBirds.R
\name{getGridIDs}
\alias{getGridIDs}
\title{Create unique IDs based on a grid}
\usage{
getGridIDs(x, grid)
}
\arguments{
\item{x}{a SpatialPointsDataFrame .}

\item{grid}{A a SpatialPolygon object (a grid is expected) defining the
maximum extent of visits effort.}
}
\value{
A vector of the same length as the number of rows (observations) as x
  with a unique number corresponding to the grid's ID.
}
\description{
Takes a spatial points dataframe and a grid and gets the overlay IDs.
}
