% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeBirds.R
\name{getGridIDs}
\alias{getGridIDs}
\title{Create unique IDs based on a grid}
\usage{
getGridIDs(x, grid, idcol)
}
\arguments{
\item{x}{a SpatialPointsDataFrame ('sp') or sf ('sf') (with the observations).}

\item{grid}{a SpatialPolygon object ('sp') or sf ('sf') (a grid is expected) defining the
maximum extent of visits effort.}

\item{idcol}{column name for the grid names or ids}
}
\value{
A vector of the same length as the number of rows (observations) as x
  with a unique number corresponding to the grid's ID.
}
\description{
Takes a spatial points data.frame and a grid and gets the overlay IDs.
}
