% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBirds.R
\name{getSpatial}
\alias{getSpatial}
\title{Collect spatial data}
\usage{
getSpatial(birdOverlay, visitCol = NULL)
}
\arguments{
\item{birdOverlay}{An birdOverlay which is the result of the \code{\link{overlayBirds}}-function.}

\item{visitCol}{A character string for specifying the columns that identify a visit.}
}
\value{
The a sf over the overlay grid with a data.frame
}
\description{
An internal function. Takes the data.frame resulting from the spatial overlay
for a specific grid and summarizes basic variables as a polygon spatial layer
}
\keyword{internal}
