% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBirds.R
\name{getSpatial}
\alias{getSpatial}
\title{Collect spatial data}
\usage{
getSpatial(birdData)
}
\arguments{
\item{birdData}{An OrganizedBirds object, which is the input to the \code{\link{overlayBirds}}-function.}

\item{visitCol}{A character string for specifying the columns that identify a visit.}
}
\value{
The a SpatialPolygonsDataFrame over the overlay gris with a dataframe
}
\description{
An internal function. Takes the dataframe resulting from the spatial overlay
for a specific grid and summarises basic variables as a polygon spatial layer
}
\keyword{internal}
