% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBirds.R
\name{getTemporal}
\alias{getTemporal}
\title{Collect temporal data}
\usage{
getTemporal(birdOverlay, visitCol = NULL)
}
\arguments{
\item{birdOverlay}{An birdOverlay which is the result of the \code{\link{overlayBirds}}-function.}

\item{visitCol}{A character string for specifying the columns that identify a visit.}
}
\value{
The overlay dataframe for a specific grid, including the spillover visits.
}
\description{
An internal function. Takes the dataframe resulting from the temporal overlay
for a specific grid and summarises basic variables as a time series
}
\keyword{internal}
