% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreVisits.r
\name{getUTMproj}
\alias{getUTMproj}
\title{A wrapper around getUTMzone and produce a proj4 string}
\usage{
getUTMproj(x)
}
\arguments{
\item{x}{an object of class \sQuote{OrganizedBirds} or \sQuote{SpatialPointsDataFrame}}
}
\value{
a proj4 character string for an appropriate UTM zone
}
\description{
A wrapper around getUTMzone and produce a proj4 string
}
\examples{
OB <- organizeBirds(bombusObs)
getUTMproj(OB)
}
