% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUTM.r
\name{getUTMzone}
\alias{getUTMzone}
\title{An internal function. Depends on the polygon object utmZones}
\usage{
getUTMzone(sf)
}
\arguments{
\item{sf}{a simple feature (sf') object. It should consist of only one geometry type}
}
\value{
a list with an element 'zone' with an integer for the UTM zone value
}
\description{
An internal function. Depends on the polygon object utmZones
}
\keyword{internal}
