% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focalSpecies.R
\name{listSpecies}
\alias{listSpecies}
\title{A function to list all species from all visits}
\usage{
listSpecies(x)
}
\arguments{
\item{x}{an object of class \sQuote{OrganizedBirds} or \sQuote{SummarizeBirds}.}
}
\value{
a \code{vector} with all species names in the data set
}
\description{
A function to list all species from all visits (total species list for data set).
}
\examples{
OB <- organizeBirds(bryophytaObs, sppCol = "scientificName", simplifySppName = TRUE)
allSpp <- listSpecies(OB)
}
\seealso{
\code{\link{summarizeBirds}}, \code{\link{exportBirds}}
}
