% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGrid.r
\name{makeCircle}
\alias{makeCircle}
\title{Create the minimum circle containing the points}
\usage{
makeCircle(spdf)
}
\arguments{
\item{spdf}{an object of class \sQuote{SpatialPointsDataFrame} with geodesic
coordinates in WGS84 (ESPG:4326).}
}
\value{
an object of class \sQuote{SpatialPolygon} with a polygon with geodesic
coordinates in WGS84 (ESPG:4326).
}
\description{
Create the minimum circle containing the points
}
