% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeBirds.R
\name{obsData}
\alias{obsData}
\alias{obsData.OrganizedBirds}
\title{Extract observation data}
\usage{
obsData(x)

\method{obsData}{OrganizedBirds}(x)
}
\arguments{
\item{x}{An OrganizedBirds-object}
}
\value{
A data.frame
}
\description{
Extract the observation data from a OrganizedBirds-object
}
\examples{
ob <- organizeBirds(bombusObs)
head(obsData(ob))
}
