% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observationIndex.R
\name{obsIndex}
\alias{obsIndex}
\title{Observation Index}
\usage{
obsIndex(x, dimension, timeRes = NULL, focalSp = NULL)
}
\arguments{
\item{x}{an object of class \sQuote{SummarizeBirds}.}

\item{dimension}{a character string indicating if the export should be
\code{"spatial"} or \code{"temporal"}}

\item{timeRes}{the time resolution as a character string if \code{dimension = "temporal"}:  \code{"Yearly", "Monthly"} or \code{"Daily"}}

\item{focalSp}{the focal species to look for.}
}
\value{
If \code{dimension = "spatial"} a \sQuote{SpatialPolygonsDataFrame} or a \sQuote{xts} timeseries if \code{dimension = "temporal"}.
}
\description{
This function extracts the proportion of observations for a focal species to
all observations (including the focal species) over time or area.
}
\examples{
grid <- makeGrid(gotaland, gridSize = 10)
PBD<-bombusObs
OB <- organizeBirds(PBD, sppCol = "scientificName", simplifySppName = TRUE)
SB <- summariseBirds(OB, grid=grid)
spp <- listSpecies(SB)
obsInd<-obsIndex(SB, "temporal", "yearly", focalSp=spp[1])
}
