% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGrid.r
\name{renameGrid}
\alias{renameGrid}
\title{Rename the cells in a grid}
\usage{
renameGrid(grid, idcol)
}
\arguments{
\item{grid}{an object of class \sQuote{sf}.}

\item{idcol}{column name with names or ids}
}
\value{
the same input object with known names
}
\description{
Takes a sf* and renames it to "ID1":"IDn".
}
\keyword{internal}
