% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeBirds.R
\name{simplifySpp}
\alias{simplifySpp}
\title{Simplify species names}
\usage{
simplifySpp(df, sppCol)
}
\arguments{
\item{df}{A data.frame with at least the column specified in sppCol}

\item{sppCol}{A character vector with the column names for the species names.}
}
\value{
A vector with data.frame with a canonical name given by taxize::gbif_parse(),
that is a scientific name with up to 3 elements and no authorship
}
\description{
Removes infraspecific epithets, authors and years from scientific names
}
