% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBirds.R
\name{summarizeBirds}
\alias{summarizeBirds}
\title{Summarize the OrganizedBirds}
\usage{
summarizeBirds(x, ...)
}
\arguments{
\item{x}{An OrgnanizedBirds-object created by \code{\link{organizeBirds}}}

\item{...}{additional parameters for \code{\link{summarizeBirds}}}
}
\value{
A SummarizedBirds-object
}
\description{
Takes a OrganizedBirds-object and a SpatialPolygons*-grid and summarizes it
in spatial and  temporal dimensions.
}
\examples{
ob<-organizeBirds(bombusObs)
grid <- makeGrid(gotaland, gridSize = 10)
SB <- summarizeBirds(ob, grid)
nObsG<-rowSums(SB$spatioTemporal[,,13,"nObs"], na.rm = FALSE)
nObsG2<-SB$spatial@data$nObs
any(nObsG != nObsG2, na.rm = TRUE) ## Check, two ways to obtain the same
}
